unit tests from network.rs moved to tests/unit/filters/network.rs
